Refactorin: divide views.py into a module with smaller easier to read files
